# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Duck Curve Home API",
    "description": "Home Energy Management System – Live-Zustand, Historie, Plan und Steuerung.",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/control/actuators/{key}": {
      "post": {
        "tags": [
          "Steuerung"
        ],
        "summary": "Aktor schalten",
        "operationId": "switch_actuator_api_v1_control_actuators__key__post",
        "parameters": [
          {
            "name": "key",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Key"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ActuatorCommandIn"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ActuatorCommandOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/control/heat-pump/mode": {
      "post": {
        "tags": [
          "Steuerung"
        ],
        "summary": "Betriebsmodus der Wärmepumpe setzen",
        "operationId": "set_mode_api_v1_control_heat_pump_mode_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/HeatPumpModeIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OperatingMode"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/control/decisions": {
      "get": {
        "tags": [
          "Steuerung"
        ],
        "summary": "Letzte Regler-Entscheidungen (Zustandswechsel)",
        "operationId": "decisions_api_v1_control_decisions_get",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 20,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Decision"
                  },
                  "title": "Response Decisions Api V1 Control Decisions Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "ActuatorCommandIn": {
        "properties": {
          "state": {
            "type": "boolean",
            "title": "State"
          },
          "duration_min": {
            "anyOf": [
              {
                "type": "integer",
                "maximum": 1440.0,
                "minimum": 1.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Duration Min"
          }
        },
        "type": "object",
        "required": [
          "state"
        ],
        "title": "ActuatorCommandIn"
      },
      "ActuatorCommandOut": {
        "properties": {
          "key": {
            "type": "string",
            "title": "Key"
          },
          "requested": {
            "type": "boolean",
            "title": "Requested"
          },
          "observed": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Observed"
          },
          "ok": {
            "type": "boolean",
            "title": "Ok"
          },
          "message_de": {
            "type": "string",
            "title": "Message De"
          }
        },
        "type": "object",
        "required": [
          "key",
          "requested",
          "observed",
          "ok",
          "message_de"
        ],
        "title": "ActuatorCommandOut"
      },
      "AutoProfile": {
        "type": "string",
        "enum": [
          "eco",
          "pv",
          "price",
          "smart"
        ],
        "title": "AutoProfile"
      },
...
      "ControllerState": {
        "type": "string",
        "enum": [
          "off",
          "idle",
          "arming",
          "released",
          "running_released",
          "cooldown",
          "manual",
          "failsafe"
        ],
        "title": "ControllerState"
      },
      "Decision": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "at": {
            "type": "string",
            "format": "date-time",
            "title": "At"
          },
          "controller_state": {
            "$ref": "#/components/schemas/ControllerState"
          },
          "k1_release": {
            "type": "boolean",
            "title": "K1 Release"
          },
          "k2_block": {
            "type": "boolean",
            "title": "K2 Block"
          },
          "reasons": {
            "items": {
              "$ref": "#/components/schemas/ReasonCode"
            },
            "type": "array",
            "title": "Reasons"
          },
          "blocked_by": {
            "items": {
              "$ref": "#/components/schemas/ReasonCode"
            },
            "type": "array",
            "title": "Blocked By"
          },
          "inputs": {
            "$ref": "#/components/schemas/DecisionInputs"
          },
          "valid_until": {
            "type": "string",
            "format": "date-time",
            "title": "Valid Until"
          },
          "next_expected": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/NextExpected"
              },
              {
                "type": "null"
              }
            ]
          },
          "explanation_de": {
            "type": "string",
            "title": "Explanation De"
          },
          "changed": {
            "type": "boolean",
            "title": "Changed",
            "default": false
          }
        },
        "type": "object",
        "required": [
          "at",
          "controller_state",
          "k1_release",
          "k2_block",
          "reasons",
          "blocked_by",
          "inputs",
          "valid_until",
          "next_expected",
          "explanation_de"
        ],
        "title": "Decision"
      },
      "DecisionInputs": {
        "properties": {
          "surplus_ewma_kw": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Surplus Ewma Kw"
          },
          "import_ewma_kw": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Import Ewma Kw"
          },
          "hp_running": {
            "type": "boolean",
            "title": "Hp Running"
          },
          "hp_power_kw": {
            "type": "number",
            "title": "Hp Power Kw"
          },
          "buffer_soc": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Buffer Soc"
          },
          "buffer_top_c": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Buffer Top C"
          },
          "price_ct_kwh": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Price Ct Kwh"
          },
          "price_rank": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Price Rank"
          },
          "outdoor_temp_c": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Outdoor Temp C"
          },
          "starts_today": {
            "type": "integer",
            "title": "Starts Today"
          },
          "seconds_since_stop": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Seconds Since Stop"
          },
          "seconds_since_start": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Seconds Since Start"
          }
        },
        "type": "object",
        "required": [
          "surplus_ewma_kw",
          "import_ewma_kw",
          "hp_running",
          "hp_power_kw",
          "buffer_soc",
          "buffer_top_c",
          "price_ct_kwh",
          "price_rank",
          "outdoor_temp_c",
          "starts_today",
          "seconds_since_stop",
          "seconds_since_start"
        ],
        "title": "DecisionInputs"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "HeatPumpModeIn": {
        "properties": {
          "system_mode": {
            "$ref": "#/components/schemas/SystemMode"
          },
          "auto_profile": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/AutoProfile"
              },
              {
                "type": "null"
              }
            ]
          },
          "manual_state": {
            "anyOf": [
              {
                "type": "string",
                "enum": [
                  "on",
                  "off"
                ]
              },
              {
                "type": "null"
              }
            ],
            "title": "Manual State"
          },
          "duration_min": {
            "type": "integer",
            "maximum": 720.0,
            "minimum": 5.0,
            "title": "Duration Min",
            "default": 120
          }
        },
        "type": "object",
        "required": [
          "system_mode"
        ],
        "title": "HeatPumpModeIn"
      },
...
      "NextExpected": {
        "properties": {
          "action": {
            "type": "string",
            "title": "Action"
          },
          "at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "At"
          },
          "because": {
            "$ref": "#/components/schemas/ReasonCode"
          },
          "text_de": {
            "type": "string",
            "title": "Text De"
          }
        },
        "type": "object",
        "required": [
          "action",
          "at",
          "because",
          "text_de"
        ],
        "title": "NextExpected"
      },
      "OperatingMode": {
        "properties": {
          "system_mode": {
            "$ref": "#/components/schemas/SystemMode",
            "default": "auto"
          },
          "auto_profile": {
            "$ref": "#/components/schemas/AutoProfile",
            "default": "smart"
          },
          "override": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Override"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "type": "object",
        "title": "OperatingMode"
      },
      "Override": {
        "properties": {
          "kind": {
            "$ref": "#/components/schemas/OverrideKind"
          },
          "started_at": {
            "type": "string",
            "format": "date-time",
            "title": "Started At"
          },
          "ends_at": {
            "type": "string",
            "format": "date-time",
            "title": "Ends At"
          },
          "set_by": {
            "type": "string",
            "title": "Set By",
            "default": "dashboard"
          }
        },
        "type": "object",
        "required": [
          "kind",
          "started_at",
          "ends_at"
        ],
        "title": "Override"
      },
      "OverrideKind": {
        "type": "string",
        "enum": [
          "force_release",
          "force_off",
          "force_block"
        ],
        "title": "OverrideKind"
      },
...
      "ReasonCode": {
        "type": "string",
        "enum": [
          "pv_surplus",
          "pv_surplus_fading",
          "price_negative",
          "price_cheap_window",
          "planned_window",
          "heat_demand_forced",
          "hp_running_own_control",
          "min_runtime_hold",
          "min_offtime_pending",
          "on_delay_pending",
          "off_delay_pending",
          "buffer_full",
          "buffer_no_headroom",
          "max_starts_reached",
          "no_trigger",
          "import_too_high",
          "manual_override",
          "mode_off",
          "sensor_stale",
          "sensor_unavailable",
          "price_data_stale",
          "hp_not_responding",
          "failsafe",
          "toggle_rate_exceeded"
        ],
        "title": "ReasonCode"
      },
...
      "SystemMode": {
        "type": "string",
        "enum": [
          "off",
          "manual",
          "auto"
        ],
        "title": "SystemMode"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}